+++ onapp2vhi/ops/cold_migrate.py
@@ -29,7 +29,14 @@ from onapp2vhi.inc.vinfra_wrapper import VinfraCommand, VinfraError
 logs = OnAppVHILogger()
 
 
-def vm_cold_migrate(cfg: OnApp2VHIConfig, vdom: str, vproj: str, idn: str, vm_properties: dict, vhi_obj, placement=''):
+def vm_cold_migrate(cfg: OnApp2VHIConfig,
+                    vdom: str,
+                    vproj: str,
+                    idn: str,
+                    vm_properties: dict,
+                    vhi_obj,
+                    placement='',
+                    cpu_hotplug=False):
     # ToDo
     #  verify IP address before running script
     if not idn:
@@ -151,7 +158,8 @@ def vm_cold_migrate(cfg: OnApp2VHIConfig, vdom: str, vproj: str, idn: str, vm_pr
                                        onapp_nics=_onapp_nics,
                                        hostname=_vm_properties['hostname'],
                                        domain=_vm_properties['domain'],
-                                       vhi_storage_policy=_vm_properties['storage_policy'])
+                                       vhi_storage_policy=_vm_properties['storage_policy'],
+                                       cpu_hotplug=cpu_hotplug)
         if not _vhi_vm_id:
             return False
     else: